Improve day 5, switch to usize totally, add everything to main

+++ aoc2021/src/day1.rs
@@ -1,10 +1,10 @@
 #[aoc_generator(day1)]
-pub fn parse_input(input: &str) -> Vec<u32> {
+pub fn parse_input(input: &str) -> Vec<usize> {
     input.lines().map(|l| l.parse().unwrap()).collect()
 }
 
 #[aoc(day1, part1)]
-pub fn solve_part1(input: &[u32]) -> u32 {
+pub fn solve_part1(input: &[usize]) -> usize {
     let mut count = 0;
     for i in 1..input.len() {
         if input[i - 1] < input[i] {
@@ -15,7 +15,7 @@ pub fn solve_part1(input: &[u32]) -> u32 {
 }
 
 #[aoc(day1, part2)]
-pub fn solve_part2(input: &[u32]) -> u32 {
+pub fn solve_part2(input: &[usize]) -> usize {
     let mut count = 0;
     for i in 3..input.len() {
         if input[i - 3] < input[i] {
@@ -29,7 +29,7 @@ pub fn solve_part2(input: &[u32]) -> u32 {
 mod tests {
     use super::*;
 
-    const EXAMPLE_INPUT: [u32; 10] = [199, 200, 208, 210, 200, 207, 240, 269, 260, 263];
+    const EXAMPLE_INPUT: [usize; 10] = [199, 200, 208, 210, 200, 207, 240, 269, 260, 263];
 
     #[test]
     fn part1_example() {
